pyproject.toml
[tool.poetry]
name = "gestão de escala"
version = "0.1.0"
description = ""
authors = ["Washington Vieira <wv817816@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.0"
sqlmodel = "^0.0.22"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
